Move to MVVM pattern

+++ app/src/main/java/ru/evgeniy/aaacourse/data/Genre.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class Genre(val id: Int, val name: String): Parcelable
+data class Genre(val id: Long, val name: String): Parcelable
